Replace git-hooks-test by centos6-perl
It downed on me that if I want to support the Perl and Git which come
with CentOS 6 the easiest way to do that is to use a CentOS 6 based
docker image!

So, I replaced the old git-hooks-test image, which was based on Ubuntu
and in which I installed from source Perl and Git, by a new image called
centos6-perl, which is based on the standard centos:6 image from Docker
Hub, on top of which I install perl and git packages. On top of that I
install App::Cpanminus from source and use it to download and install
all of the dependencies of my Git::Hooks module.

So, the new image is in good shape to become the testing base for my
Perl modules.